fix(rebase): resolve conflicts
