+++ cpp/src/arrow/engine/substrait/test_util.h
@@ -26,7 +26,7 @@
 // global namespace conflicts.
 
 #include "arrow/compute/exec/exec_plan.h"
-#include "arrow/compute/exec/expression_internal.h"
+#include "arrow/compute/exec/util.h"
 #include "arrow/dataset/file_base.h"
 #include "arrow/engine/substrait/util.h"
 #include "arrow/table.h"
@@ -36,69 +36,10 @@ namespace arrow {
 
 namespace engine {
 
-class NullSinkNodeConsumer : public compute::SinkNodeConsumer {
- public:
-  Status Init(const std::shared_ptr<Schema>&, compute::BackpressureControl*) override {
-    return Status::OK();
-  }
-  Status Consume(compute::ExecBatch exec_batch) override { return Status::OK(); }
-  Future<> Finish() override { return Status::OK(); }
-
- public:
-  static std::shared_ptr<NullSinkNodeConsumer> Make() {
-    return std::make_shared<NullSinkNodeConsumer>();
-  }
-};
-
-const auto kNullConsumer = std::make_shared<NullSinkNodeConsumer>();
-
-const std::shared_ptr<Schema> kBoringSchema = schema({
-    field("bool", boolean()),
-    field("i8", int8()),
-    field("i32", int32()),
-    field("i32_req", int32(), /*nullable=*/false),
-    field("u32", uint32()),
-    field("i64", int64()),
-    field("f32", float32()),
-    field("f32_req", float32(), /*nullable=*/false),
-    field("f64", float64()),
-    field("date64", date64()),
-    field("str", utf8()),
-    field("list_i32", list(int32())),
-    field("struct", struct_({
-                        field("i32", int32()),
-                        field("str", utf8()),
-                        field("struct_i32_str",
-                              struct_({field("i32", int32()), field("str", utf8())})),
-                    })),
-    field("list_struct", list(struct_({
-                             field("i32", int32()),
-                             field("str", utf8()),
-                             field("struct_i32_str", struct_({field("i32", int32()),
-                                                              field("str", utf8())})),
-                         }))),
-    field("dict_str", dictionary(int32(), utf8())),
-    field("dict_i32", dictionary(int32(), int32())),
-    field("ts_ns", timestamp(TimeUnit::NANO)),
-});
+const auto kNullConsumer = std::make_shared<compute::NullSinkNodeConsumer>();
 
 ARROW_ENGINE_EXPORT std::shared_ptr<DataType> StripFieldNames(
     std::shared_ptr<DataType> type);
-
-ARROW_ENGINE_EXPORT inline compute::Expression UseBoringRefs(
-    const compute::Expression& expr) {
-  if (expr.literal()) return expr;
-
-  if (auto ref = expr.field_ref()) {
-    return compute::field_ref(*ref->FindOne(*kBoringSchema));
-  }
-
-  auto modified_call = *CallNotNull(expr);
-  for (auto& arg : modified_call.arguments) {
-    arg = UseBoringRefs(arg);
-  }
-  return compute::Expression{std::move(modified_call)};
-}
 
 ARROW_ENGINE_EXPORT void WriteIpcData(const std::string& path,
                                       const std::shared_ptr<fs::FileSystem> file_system,
@@ -109,9 +50,8 @@ ARROW_ENGINE_EXPORT Result<std::shared_ptr<Table>> GetTableFromPlan(
     const std::shared_ptr<Schema>& output_schema);
 
 ARROW_ENGINE_EXPORT void CheckRoundTripResult(
-    const std::shared_ptr<Schema> output_schema,
-    const std::shared_ptr<Table> expected_table, compute::ExecContext& exec_context,
-    std::shared_ptr<Buffer>& buf, const std::vector<int>& include_columns = {},
+    const std::shared_ptr<Table> expected_table, std::shared_ptr<Buffer>& buf,
+    const std::vector<int>& include_columns = {},
     const ConversionOptions& conversion_options = {},
     const compute::SortOptions* sort_options = NULLPTR);
